+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Fix formatting of stdout and stderr during command execution
+
 ## [0.3.0] - 2016-10-11
 ### Added
 - Include stdout in the exception raised when ExecCommand fails